+++ src/app/src/components/FilterSidebarSearchTab.jsx
@@ -13,7 +13,6 @@ import Tooltip from '@material-ui/core/Tooltip';
 import ReactSelect from 'react-select';
 import get from 'lodash/get';
 
-import FacilitySidebarSearchTabFacilitiesCount from './FacilitySidebarSearchTabFacilitiesCount';
 import ShowOnly from './ShowOnly';
 
 import {
@@ -22,12 +21,16 @@ import {
     updateContributorTypeFilter,
     updateCountryFilter,
     updateCombineContributorsFilterOption,
+    updateBoundaryFilter,
     resetAllFilters,
 } from '../actions/filters';
 
 import { fetchFacilities } from '../actions/facilities';
 
-import { recordSearchTabResetButtonClick } from '../actions/ui';
+import {
+    recordSearchTabResetButtonClick,
+    showDrawFilter,
+} from '../actions/ui';
 
 import {
     contributorOptionsPropType,
@@ -85,6 +88,9 @@ function FilterSidebarSearchTab({
     fetchingOptions,
     submitFormOnEnterKeyPress,
     vectorTileFlagIsActive,
+    activateDrawFilter,
+    clearDrawFilter,
+    boundary,
 }) {
     if (fetchingOptions) {
         return (
@@ -156,6 +162,28 @@ function FilterSidebarSearchTab({
                 </li>
             </ol>
         </div>
+    );
+
+    const boundaryButton = boundary == null ? (
+        <Button
+            variant="outlined"
+            onClick={activateDrawFilter}
+            disableRipple
+            color="primary"
+            className="outlined-button outlined-button--full-width"
+        >
+            DRAW AREA
+        </Button>
+    ) : (
+        <Button
+            variant="outlined"
+            onClick={clearDrawFilter}
+            disableRipple
+            color="primary"
+            className="outlined-button outlined-button--full-width"
+        >
+            REMOVE AREA
+        </Button>
     );
 
     return (
@@ -258,6 +286,16 @@ function FilterSidebarSearchTab({
                         disabled={fetchingOptions || fetchingFacilities}
                     />
                 </div>
+                <div className="form__field">
+                    <InputLabel
+                        shrink={false}
+                        htmlFor={CONTRIBUTORS}
+                        style={filterSidebarSearchTabStyles.inputLabelStyle}
+                    >
+                        Filter by Area
+                    </InputLabel>
+                    {boundaryButton}
+                </div>
                 <div className="form__action">
                     <a
                         className="control-link"
@@ -302,7 +340,6 @@ function FilterSidebarSearchTab({
                 </div>
                 {noFacilitiesFoundMessage}
             </div>
-            <FacilitySidebarSearchTabFacilitiesCount />
         </div>
     );
 }
@@ -354,6 +391,7 @@ function mapStateToProps({
         contributorTypes,
         countries,
         combineContributors,
+        boundary,
     },
     facilities: {
         facilities: {
@@ -377,6 +415,7 @@ function mapStateToProps({
         combineContributors,
         fetchingFacilities,
         facilities,
+        boundary,
         fetchingOptions: fetchingContributors
             || fetchingContributorTypes
             || fetchingCountries,
@@ -413,6 +452,12 @@ function mapDispatchToProps(dispatch, {
         submitFormOnEnterKeyPress: makeSubmitFormOnEnterKeyPressFunction(
             () => dispatch(fetchFacilities(push)),
         ),
+        activateDrawFilter: () => dispatch(showDrawFilter(true)),
+        clearDrawFilter: () => {
+            dispatch(showDrawFilter(false));
+            dispatch(updateBoundaryFilter(null));
+            return dispatch(fetchFacilities({}));
+        },
     };
 }