+++ src/django/api/tests.py
@@ -27,7 +27,8 @@ from api.models import (Facility, FacilityList, FacilityListItem,
 from api.oar_id import make_oar_id, validate_oar_id
 from api.matching import match_facility_list_items
 from api.processing import (parse_facility_list_item,
-                            geocode_facility_list_item)
+                            geocode_facility_list_item,
+                            reduce_matches)
 from api.geocoding import (create_geocoding_params,
                            format_geocoded_address_data,
                            geocode_address)
@@ -1189,6 +1190,18 @@ class DedupeMatchingTests(TestCase):
         result = match_facility_list_items(facility_list)
         self.assertTrue(result['results']['no_geocoded_items'])
 
+    def test_reduce_matches(self):
+        matches = [
+            ('US2020052GKF19F', 75),
+            ('US2020052GKF19F_MATCH-23', 88),
+            ('US2020052YDVKBQ', 45)
+        ]
+        expected = [
+            ('US2020052GKF19F', 88),
+            ('US2020052YDVKBQ', 45)
+        ]
+        self.assertEqual(expected, reduce_matches(matches))
+
 
 class OarIdTests(TestCase):
 
@@ -1223,22 +1236,26 @@ class ContributorsListAPIEndpointTests(TestCase):
         self.email_two = 'two@example.com'
         self.email_three = 'three@example.com'
         self.email_four = 'four@example.com'
+        self.email_five = 'five@example.com'
 
         self.contrib_one_name = 'contributor that should be included'
         self.contrib_two_name = 'contributor with no lists'
         self.contrib_three_name = 'contributor with an inactive list'
         self.contrib_four_name = 'contributor with a non public list'
+        self.contrib_five_name = 'contributor with only error items'
 
         self.country_code = 'US'
         self.list_one_name = 'one'
         self.list_one_b_name = 'one-b'
         self.list_three_name = 'three'
         self.list_four_name = 'four'
+        self.list_five_name = 'five'
 
         self.user_one = User.objects.create(email=self.email_one)
         self.user_two = User.objects.create(email=self.email_two)
         self.user_three = User.objects.create(email=self.email_three)
         self.user_four = User.objects.create(email=self.email_four)
+        self.user_five = User.objects.create(email=self.email_five)
 
         self.contrib_one = Contributor \
             .objects \
@@ -1262,6 +1279,12 @@ class ContributorsListAPIEndpointTests(TestCase):
             .objects \
             .create(admin=self.user_four,
                     name=self.contrib_four_name,
+                    contrib_type=Contributor.OTHER_CONTRIB_TYPE)
+
+        self.contrib_five = Contributor \
+            .objects \
+            .create(admin=self.user_five,
+                    name=self.contrib_five_name,
                     contrib_type=Contributor.OTHER_CONTRIB_TYPE)
 
         self.list_one = FacilityList \
@@ -1318,6 +1341,24 @@ class ContributorsListAPIEndpointTests(TestCase):
                     is_active=True,
                     contributor=self.contrib_four)
 
+        self.list_five = FacilityList \
+            .objects \
+            .create(header="header",
+                    file_name="one",
+                    name=self.list_five_name)
+
+        self.source_five = Source \
+            .objects \
+            .create(source_type=Source.LIST,
+                    facility_list=self.list_five,
+                    contributor=self.contrib_five)
+
+        self.list_item_five = FacilityListItem \
+            .objects \
+            .create(row_index=0,
+                    source=self.source_five,
+                    status=FacilityListItem.ERROR_PARSING)
+
     def test_contributors_list_has_only_contributors_with_active_lists(self):
         response = self.client.get('/api/contributors/')
         response_data = response.json()
@@ -1340,6 +1381,11 @@ class ContributorsListAPIEndpointTests(TestCase):
 
         self.assertNotIn(
             self.contrib_four_name,
+            contributor_names,
+        )
+
+        self.assertNotIn(
+            self.contrib_five_name,
             contributor_names,
         )
 
@@ -3840,7 +3886,7 @@ class SerializeOtherLocationsTest(FacilityAPITestCaseBase):
         data = json.loads(response.content)
         self.assertEqual(
             len(data['properties']['other_locations']),
-            2,
+            3,
         )
 
         self.assertEqual(
@@ -3874,9 +3920,10 @@ class SerializeOtherLocationsTest(FacilityAPITestCaseBase):
         )
 
         data = json.loads(response.content)
+
         self.assertEqual(
             len(data['properties']['other_locations']),
-            2,
+            3,
         )
 
         self.assertEqual(